update to wrapper

+++ scripts/etlwrapper.sh
@@ -33,57 +33,59 @@ mkdir hierarchies
 mkdir processing
 echo "########"
 ## build or update mapping file
-echo ""
-echo "#### Making mappings and update job config for current run ####"
-echo ""
+
 
 aws s3 cp s3://avillach-73-bdcatalyst-etl/general/data/ data/ --recursive
 
 cp ETL-MissionControl-dbgap-submodule/jars/* .
 
-for studyid in ${studyids[@]}; do
+if [ $1 = "true"]; then
+   echo ""
+   echo "#### Making mappings and update job config for current run ####"
+   echo ""
+   for studyid in ${studyids[@]}; do
 
-   find data/ -name "phs*" -exec rm -rf {} \;
+      find data/ -name "phs*" -exec rm -rf {} \;
 
-   rm -rf completed/*
-   mkdir data
+      rm -rf completed/*
+      mkdir data
 
-   aws s3 cp s3://avillach-73-bdcatalyst-etl/${studyid}/mappings/ mappings/ --recursive --include "mapping.csv" --quiet
+      aws s3 cp s3://avillach-73-bdcatalyst-etl/${studyid}/mappings/ mappings/ --recursive --include "mapping.csv" --quiet
 
-   aws s3 cp s3://avillach-73-bdcatalyst-etl/${studyid}/resources/ resources/ --recursive --include "job.config" --quiet
+      aws s3 cp s3://avillach-73-bdcatalyst-etl/${studyid}/resources/ resources/ --recursive --include "job.config" --quiet
 
-   sed -i "s/skipdataheader=Y/skipdataheader=N/g" resources/job.config
+      sed -i "s/skipdataheader=Y/skipdataheader=N/g" resources/job.config
 
-   sed -i "s/patientmappingfile=.*/patientmappingfile=data\/${studyid^^}\_PatientMapping.csv/g" resources/job.config
+      sed -i "s/patientmappingfile=.*/patientmappingfile=data\/${studyid^^}\_PatientMapping.csv/g" resources/job.config
 
-   echo 'usepatientmapping=Y' >> resources/job.config
+      echo 'usepatientmapping=Y' >> resources/job.config
 
-   aws s3 cp s3://avillach-73-bdcatalyst-etl/${studyid}/rawData/data/ data/ --recursive --quiet
+      aws s3 cp s3://avillach-73-bdcatalyst-etl/${studyid}/rawData/data/ data/ --recursive --quiet
 
-   java -jar DbgapDecodeFiles.jar
+      java -jar DbgapDecodeFiles.jar
 
-   aws s3 cp completed/ s3://avillach-73-bdcatalyst-etl/${studyid}/data/ --recursive --quiet
+      aws s3 cp completed/ s3://avillach-73-bdcatalyst-etl/${studyid}/data/ --recursive --quiet
 
-   java -jar DbgapTreeBuilder2.jar -dataseperator '\t'
+      java -jar DbgapTreeBuilder2.jar -dataseperator '\t'
 
-   find data/ -name "phs*" -exec rm -rf {} \;
+      find data/ -name "phs*" -exec rm -rf {} \;
 
-   mv completed/* data/
+      mv completed/* data/
 
-   java -jar DataAnalyzer.jar -propertiesfile resources/job.config
+      java -jar DbGapPMGenerator.jar -propertiesfile resources/job.config
 
-   java -jar DbGapPMGenerator.jar -propertiesfile resources/job.config
+      aws s3 cp completed/${studyid^^}_PatientMapping.csv s3://avillach-73-bdcatalyst-etl/${studyid}/data/
 
-   aws s3 cp completed/${studyid^^}_PatientMapping.csv s3://avillach-73-bdcatalyst-etl/${studyid}/data/
+      aws s3 cp mappings/mapping.csv s3://avillach-73-bdcatalyst-etl/${studyid}/currentmapping.csv --quiet
 
-   aws s3 cp mappings/mapping.csv s3://avillach-73-bdcatalyst-etl/${studyid}/currentmapping.csv --quiet
+      aws s3 cp resources/job.config s3://avillach-73-bdcatalyst-etl/${studyid}/current.config --quiet
+      
+   done
 
-   aws s3 cp resources/job.config s3://avillach-73-bdcatalyst-etl/${studyid}/current.config --quiet
-   
-done
-echo ""
-echo "#### Finished building mappings and job.config for current run ####"
-echo ""
+   echo ""
+   echo "#### Finished building mappings and job.config for current run ####"
+   echo ""
+fi
 
 echo ""
 echo "#### Building allConcepts.csv for each study ####"
@@ -119,13 +121,25 @@ for studyid in ${studyids[@]}; do
 
    #java -jar GenerateAllConcepts.jar -propertiesfile resources/job.config
    
-   python runpartition2.py
+   #ython runpartition2.py
+   echo ""
+   echo "#### Building partitioned files ####"
+   echo ""
+   bash runpartition.sh -j 4 -m 3g -c 'config.part*.config' -r resources/
 
+   echo ""
+   echo "#### Merging partitioned files ####"
+   echo ""
    java -jar MergePartitions.jar -propertiesfile resources/job.config
 
    aws s3 cp completed/ s3://avillach-73-bdcatalyst-etl/${studyid}/completed/ --recursive --quiet
    #aws s3 cp completed/ s3://avillach-73-bdcatalyst-etl/${studyid}/completed/ --recursive
 
+   echo ""
+   echo "#### Finished Building " ${studyid} " allConcepts.csv ####"
+   echo ""
+   echo ""
+   echo ""
 done
 
 echo ""
